Enum ncollide3d::bounding_volume::PolyhedralCone

pub enum PolyhedralCone<N: Real> {
    Full,
    Empty,
    HalfLine(Unit<Vector<N>>),
    HalfSpace(Unit<Vector<N>>),
    OrthogonalSubspace(Unit<Vector<N>>),
    Span(SmallVec<[Unit<Vector<N>>; 4]>),
}

A convex cone with polyhedral faces and its apex at the origin.

A polyhedral cone is a set of half-lines forming a convex set. It is usually used to bound a set of directions like normals and tangents. It must be convex and can be generated from a finite set of vectors.

Variants
Full

A polyhedral cone which is the whole space.

Empty

An empty cone containing only the zero vector.

HalfLine(Unit<Vector<N>>)

The half-line starting at the origin, pointing toward the given diretion.

HalfSpace(Unit<Vector<N>>)

The half-space which boundary has the given diretion as normal.

OrthogonalSubspace(Unit<Vector<N>>)

The subspace orthogonal to the given diretion.

Span(SmallVec<[Unit<Vector<N>>; 4]>)

All the positive linear combinations of the given set of vectors.

Methods

impl<N: Real> PolyhedralCone<N>

pub fn unwrap_half_line(&self) -> Unit<Vector<N>>

If this polyhedral cone spans a single half-line, returns its direction.

pub fn transform_by(&mut self, m: &Isometry<N>)

Applies the given transformation to each direction bounded by this cone.

pub fn contains(&self, v: &Vector<N>) -> bool

Tests whether the given vector is contained by this cone.

pub fn polar_contains_dir(&self, dir: &Unit<Vector<N>>) -> bool

Tests if the polar of this cone contains the given direction.

This test is much sheaper than .contains().

pub fn contains_dir(&self, dir: &Unit<Vector<N>>) -> bool

Tests if this cone contains the given unit direction.