Struct ncollide3d::bounding_volume::AABB

pub struct AABB<N: Real> { /* fields omitted */ }

An Axis Aligned Bounding Box.

Methods

impl<N: Real> AABB<N>

pub fn new(mins: Point<N>, maxs: Point<N>) -> AABB<N>

Creates a new AABB.

Arguments:

mins - position of the point with the smallest coordinates.

maxs - position of the point with the highest coordinates. Each component of mins must be smaller than the related components of maxs.

pub fn mins(&self) -> &Point<N>

Reference to the AABB point with the smallest components along each axis.

pub fn maxs(&self) -> &Point<N>

Reference to the AABB point with the biggest components along each axis.

pub fn center(&self) -> Point<N>

The center of this AABB.

pub fn half_extents(&self) -> Vector<N>

The half extents of this AABB.