Struct ncollide3d::broad_phase::BroadPhasePairFilters

pub struct BroadPhasePairFilters<N: Real, T> { /* fields omitted */ }

Filters deciding whether a proximity is to be further investigated by the narrow phase or not.

All filters have have to return true in order to allow a proximity to be further handled.

Methods

impl<N: Real, T> BroadPhasePairFilters<N, T>

pub fn new() -> BroadPhasePairFilters<N, T>

Creates a new set of collision filters.

pub fn register_collision_filter(
&mut self,
name: &str,
callback: Box<BroadPhasePairFilter<N, T>>
)

Registers a collision filter.

pub fn unregister_collision_filter(&mut self, name: &str) -> bool

Unregisters a collision filter.

Returns true if the filter was found.

pub fn is_pair_valid(
&self,
b1: &CollisionObject<N, T>,
b2: &CollisionObject<N, T>
) -> bool

Tells if the collision between b1 and b2 is to be handled by the narrow-phase.