Struct ncollide3d::broad_phase::DBVTBroadPhase

pub struct DBVTBroadPhase<N: Real, BV, T> { /* fields omitted */ }

Broad phase based on a Dynamic Bounding Volume Tree.

It uses two separate trees: one for static objects and which is never updated, and one for moving objects.

Methods

impl<N, BV, T> DBVTBroadPhase<N, BV, T> where
N: Real,
BV: 'static + BoundingVolume<N> + Clone,

pub fn new(margin: N) -> DBVTBroadPhase<N, BV, T>

Creates a new broad phase based on a Dynamic Bounding Volume Tree.

pub fn num_interferences(&self) -> usize

Number of interferences detected by this broad phase.