Trait ncollide3d::interpolation::RigidMotion

pub trait RigidMotion<N: RealField + Copy> {
    fn position_at_time(&self, t: N) -> Isometry<N>;
}

A continuous rigid motion.

This is a function, assumed to be continuous, that, given a parameter t returns a direct isometry. Mathematically speaking this is a one-parameter curve on the space of direct isometries. This curve should have a continuity of at least C0.

Required Methods

fn position_at_time(&self, t: N) -> Isometry<N>

Get a position at the time t.