Module ncollide3d::narrow_phase

Persistent collision detection algorithms to compute contact points.

Structs

BallBallManifoldGenerator
Collision detector between two balls.

BallBallProximityDetector
Proximity detector between two balls.

BallConvexPolyhedronManifoldGenerator
Collision detector between two balls.

CompositeShapeShapeManifoldGenerator
Collision detector between a concave shape and another shape.

CompositeShapeShapeProximityDetector
Proximity detector between a concave shape and another shape.

ContactManifolds
An iterator through contact manifolds.

ContactPairs
Iterator through contact pairs.

ConvexPolyhedronConvexPolyhedronManifoldGenerator
Persistent contact manifold computation between two shapes having a support mapping function.

DefaultContactDispatcher
Collision dispatcher for shapes defined by ncollide_entities.

DefaultNarrowPhase
Collision detector dispatcher for collision objects.

DefaultProximityDispatcher
Proximity dispatcher for shapes defined by ncollide_entities.

PlaneBallManifoldGenerator
Collision detector between g1 plane and g1 shape implementing the SupportMap trait.

PlaneConvexPolyhedronManifoldGenerator
Collision detector between g1 plane and g1 shape implementing the SupportMap trait.

PlaneSupportMapProximityDetector
Proximity detector between a plane and a shape implementing the SupportMap trait.

ProximityPairs
Iterator through proximity pairs.

SupportMapPlaneProximityDetector
Proximity detector between a plane and a shape implementing the SupportMap trait.

SupportMapSupportMapProximityDetector
Persistent proximity detector between two shapes having a support mapping function.

Traits

ContactDispatcher

ContactManifoldGenerator
Trait implemented algorithms that compute contact points, normals and penetration depths.

NarrowPhase
Trait implemented by the narrow phase manager.

ProximityDetector
Trait implemented by algorithms that determine if two objects are in close proximity.

ProximityDispatcher

Type Definitions

ContactAlgorithm

ProximityAlgorithm