Struct ncollide3d::narrow_phase::CompositeShapeShapeManifoldGenerator

pub struct CompositeShapeShapeManifoldGenerator<N> { /* fields omitted */ }

Collision detector between a concave shape and another shape.

Methods

impl<N> CompositeShapeShapeManifoldGenerator<N>

pub fn new(flip: bool) -> CompositeShapeShapeManifoldGenerator<N>

Creates a new collision detector between a concave shape and another shape.