Trait ncollide3d::narrow_phase::NarrowPhase

pub trait NarrowPhase<N: Real, T>: Any + Send + Sync {
    fn update(
        &mut self, 
        objects: &CollisionObjectSlab<N, T>, 
        contact_events: &mut ContactEvents, 
        proximity_events: &mut ProximityEvents, 
        timestamp: usize
    );

    fn handle_interaction(
        &mut self, 
        contact_signal: &mut ContactEvents, 
        proximity_signal: &mut ProximityEvents, 
        objects: &CollisionObjectSlab<N, T>, 
        handle1: CollisionObjectHandle, 
        handle2: CollisionObjectHandle, 
        started: bool
    );

    fn handle_removal(
        &mut self, 
        objects: &CollisionObjectSlab<N, T>, 
        handle1: CollisionObjectHandle, 
        handle2: CollisionObjectHandle
    );

    fn contact_pairs<'a>(
        &'a self, 
        objects: &'a CollisionObjectSlab<N, T>
    ) -> ContactPairs<'a, N, T>;

    fn proximity_pairs<'a>(
        &'a self, 
        objects: &'a CollisionObjectSlab<N, T>
    ) -> ProximityPairs<'a, N, T>;
}

Trait implemented by the narrow phase manager.

The narrow phase manager is responsible for creating, updating and generating contact pairs between objects identified by the broad phase.

Required Methods

fn update(
&mut self,
objects: &CollisionObjectSlab<N, T>,
contact_events: &mut ContactEvents,
proximity_events: &mut ProximityEvents,
timestamp: usize
)

Updates this narrow phase.

fn handle_interaction(
&mut self,
contact_signal: &mut ContactEvents,
proximity_signal: &mut ProximityEvents,
objects: &CollisionObjectSlab<N, T>,
handle1: CollisionObjectHandle,
handle2: CollisionObjectHandle,
started: bool
)

Called when the broad phase detects that two objects are, or stop to be, in close proximity.

fn handle_removal(
&mut self,
objects: &CollisionObjectSlab<N, T>,
handle1: CollisionObjectHandle,
handle2: CollisionObjectHandle
)

Called when the interactions between two objects have to be removed because at least one of the objects is being removed.

While either objects[handle1] or objects[handle2] is being removed, the handle_removal is assumed to be called before the removal from the list objects is done.

Important traits for ContactPairs<'a, N, T>
impl<'a, N: Real, T> Iterator for ContactPairs<'a, N, T> type Item = (&'a CollisionObject<N, T>, &'a CollisionObject<N, T>, &'a ContactAlgorithm<N>);

fn contact_pairs<'a>(
&'a self,
objects: &'a CollisionObjectSlab<N, T>
) -> ContactPairs<'a, N, T>

Returns all the potential contact pairs found during the broad phase, and validated by the narrow phase.

Important traits for ProximityPairs<'a, N, T>
impl<'a, N: Real, T> Iterator for ProximityPairs<'a, N, T> type Item = (&'a CollisionObject<N, T>, &'a CollisionObject<N, T>, &'a ProximityAlgorithm<N>);

fn proximity_pairs<'a>(
&'a self,
objects: &'a CollisionObjectSlab<N, T>
) -> ProximityPairs<'a, N, T>

Returns all the potential proximity pairs found during the broad phase, and validated by the narrow phase.