Struct ncollide3d::partitioning::BVT

pub struct BVT<B, BV> { /* fields omitted */ }

A Bounding Volume Tree.

Methods

impl<B, BV> BVT<B, BV>

pub fn new_with_partitioner<F: FnMut(usize, Vec<(B, BV)>) -> (BV, BinaryPartition<B, BV>)>(
leaves: Vec<(B, BV)>,
partitioner: &mut F
) -> BVT<B, BV>

Builds a bounding volume tree using an user-defined construction function.

pub fn visit<Vis: BVTVisitor<B, BV>>(&self, visitor: &mut Vis)

Traverses this tree using an object implementing the BVTVisitortrait.

This will traverse the whole tree and call the visitor .visit_internal(...) (resp. .visit_leaf(...)) method on each internal (resp. leaf) node.

pub fn visit_bvtt<Vis: BVTTVisitor<B, BV>>(
&self,
other: &BVT<B, BV>,
visitor: &mut Vis
)

Visits the bounding volume traversal tree implicitly formed with other.

pub fn best_first_search<'a, N, BFS>(
&'a self,
algorithm: &mut BFS
) -> Option<(&'a B, BFS::UserData)> where
N: Real,
BFS: BVTCostFn<N, B, BV>,

Performs a best-fist-search on the tree.

Returns the content of the leaf with the smallest associated cost, and a result of user-defined type.

pub fn root_bounding_volume<'r>(&'r self) -> Option<&'r BV>

Reference to the bounding volume of the tree root.

pub fn depth(&self) -> usize

Computes the depth of this tree.

impl<B, BV> BVT<B, BV>

pub fn new_balanced<N>(leaves: Vec<(B, BV)>) -> BVT<B, BV> where
N: Real,
BV: BoundingVolume<N> + Clone,

Creates a balanced BVT.

pub fn median_partitioner_with_centers<N, F: FnMut(&B, &BV) -> Point<N>>(
depth: usize,
leaves: Vec<(B, BV)>,
center: &mut F
) -> (BV, BinaryPartition<B, BV>) where
N: Real,
BV: BoundingVolume<N> + Clone,

Construction function for a kdree to be used with BVT::new_with_partitioner.

pub fn median_partitioner<N>(
depth: usize,
leaves: Vec<(B, BV)>
) -> (BV, BinaryPartition<B, BV>) where
N: Real,
BV: BoundingVolume<N> + Clone,

Construction function for a kdree to be used with BVT::new_with_partitioner.