Struct ncollide3d::partitioning::DBVT

pub struct DBVT<N: Real, B, BV> { /* fields omitted */ }

A boundin volume hierarchy on which objects can be added or removed after construction.

Methods

impl<N: Real, B, BV: BoundingVolume<N>> DBVT<N, B, BV>

pub fn new() -> DBVT<N, B, BV>

Creates a new empty dynamic bonding volume hierarchy.

pub fn is_empty(&self) -> bool

Indicates whether this DBVT empty.

pub fn insert(&mut self, leaf: DBVTLeaf<N, B, BV>) -> DBVTLeafId

Inserts a leaf into this DBVT.

pub fn remove(&mut self, leaf_id: DBVTLeafId) -> DBVTLeaf<N, B, BV>

Removes a leaf from this DBVT.

Panics if the provided leaf is not attached to this DBVT.

pub fn visit<Vis: BVTVisitor<B, BV>>(&self, visitor: &mut Vis)

Traverses this tree using an object implementing the BVTVisitortrait.

This will traverse the whole tree and call the visitor .visit_internal(...) (resp. .visit_leaf(...)) method on each internal (resp. leaf) node.