ncollide3d::partitioning

Trait SimultaneousVisitor

pub trait SimultaneousVisitor<T, BV> {
    // Required method
    fn visit(
        &mut self,
        left_bv: &BV,
        left_data: Option<&T>,
        right_bv: &BV,
        right_data: Option<&T>,
    ) -> VisitStatus;
}

Trait implemented by visitor called during a simultaneous spatial partitioning data structure tarversal.

Required Methods

fn visit( &mut self, left_bv: &BV, left_data: Option<&T>, right_bv: &BV, right_data: Option<&T>, ) -> VisitStatus

Execute an operation on the content of two nodes, one from each structure.

Returns whether the traversal should continue on the nodes children, if it should not continue on those children, or if the whole traversal should be exited early.