Function ncollide3d::pipeline::glue::first_interference_with_ray

pub fn first_interference_with_ray<'a, 'b, N: RealField, Objects: CollisionObjectSet<N>>(
    objects: &'a Objects, 
    broad_phase: &'a impl BroadPhase<N, AABB<N>, Objects::CollisionObjectHandle> + ?Sized, 
    ray: &'b Ray<N>, 
    max_toi: N, 
    groups: &'b CollisionGroups
) -> Option<FirstInterferenceWithRay<'a, N, Objects>>

Returns an the closest collision object intersecting with the given ray.

The result will only include collision objects in a group that can interact with the given groups.