Function ncollide3d::pipeline::glue::interferences_with_aabb

pub fn interferences_with_aabb<'a, 'b, N, Objects>(
    objects: &'a Objects,
    broad_phase: &'a (impl BroadPhase<N, AABB<N>, Objects::CollisionObjectHandle> + ?Sized),
    aabb: &AABB<N>,
    groups: &'b CollisionGroups
) -> InterferencesWithAABB<'a, 'b, N, Objects> 
where
    N: RealField + Copy,
    Objects: CollisionObjectSet<N>,

Returns an iterator yielding all the collision objects with an AABB intersecting with the given AABB.

The result will only include collision objects in a group that can interact with the given groups.