ncollide3d::pipeline::glue

Function interferences_with_point

pub fn interferences_with_point<'a, 'b, N, Objects>(
    objects: &'a Objects,
    broad_phase: &'a (impl BroadPhase<N, AABB<N>, Objects::CollisionObjectHandle> + ?Sized),
    point: &'b Point<N>,
    groups: &'b CollisionGroups,
) -> InterferencesWithPoint<'a, 'b, N, Objects> 
where
    N: RealField + Copy,
    Objects: CollisionObjectSet<N>,

Returns an iterator yielding all the collision objects containing the given point.

The result will only include collision objects in a group that can interact with the given groups.