Function ncollide3d::pipeline::glue::interferences_with_ray

pub fn interferences_with_ray<'a, 'b, N, Objects>(
    objects: &'a Objects,
    broad_phase: &'a (impl BroadPhase<N, AABB<N>, Objects::CollisionObjectHandle> + ?Sized),
    ray: &'b Ray<N>,
    max_toi: N,
    groups: &'b CollisionGroups
) -> InterferencesWithRay<'a, 'b, N, Objects> 
where
    N: RealField + Copy,
    Objects: CollisionObjectSet<N>,

Returns an iterator yielding all the collision objects intersecting with the given ray.

The result will only include collision objects in a group that can interact with the given groups.