ncollide3d::pipeline::narrow_phase

Struct CapsuleShapeManifoldGenerator

pub struct CapsuleShapeManifoldGenerator<N: RealField + Copy> { /* private fields */ }

Collision detector between a concave shape and another shape.

Implementations

impl<N: RealField + Copy> CapsuleShapeManifoldGenerator<N>

pub fn new(flip: bool) -> CapsuleShapeManifoldGenerator<N>

Creates a new collision detector between a concave shape and another shape.