ncollide3d::pipeline::object

Struct CollisionObjectSlab

pub struct CollisionObjectSlab<N: RealField + Copy, T> { /* private fields */ }

A set of collision objects that can be indexed by collision object handles.

Implementations

impl<N: RealField + Copy, T> CollisionObjectSlab<N, T>

pub fn new() -> CollisionObjectSlab<N, T>

Creates a new empty collection of collision objects.

pub fn with_capacity(capacity: usize) -> CollisionObjectSlab<N, T>

Constructs a new empty collection with the specified capacity.

pub fn insert(&mut self, co: CollisionObject<N, T>) -> CollisionObjectSlabHandle

Inserts a new collision object into this collection and returns the corresponding handle.

pub fn remove( &mut self, handle: CollisionObjectSlabHandle, ) -> CollisionObject<N, T>

Removes from this collection the collision object identified by the given handle.

The removed collision object structure is returned.

pub fn get( &self, handle: CollisionObjectSlabHandle, ) -> Option<&CollisionObject<N, T>>

If it exists, retrieves a reference to the collision object identified by the given handle.

pub fn get_mut( &mut self, handle: CollisionObjectSlabHandle, ) -> Option<&mut CollisionObject<N, T>>

If it exists, retrieves a mutable reference to the collision object identified by the given handle.

pub fn get_pair_mut( &mut self, handle1: CollisionObjectSlabHandle, handle2: CollisionObjectSlabHandle, ) -> (Option<&mut CollisionObject<N, T>>, Option<&mut CollisionObject<N, T>>)

If they exists, retrieves a mutable reference to the two collision object identified by the given handles.

Panics if both handles are equal.

pub fn contains(&self, handle: CollisionObjectSlabHandle) -> bool

Returns true if the specified handle identifies a collision object stored in this collection.

pub fn iter(&self) -> CollisionObjects<'_, N, T>

Retrieves an iterator yielding references to each collision object.

pub fn iter_mut(&mut self) -> CollisionObjectsMut<'_, N, T>

Retrieves an iterator yielding references to each collision object.

pub fn len(&self) -> usize

The number of collision objects on this slab.

pub fn capacity(&self) -> usize

Return the number of values the slab can store without reallocating.

pub fn reserve(&mut self, additional: usize)

Reserve capacity for at least additional more values to be stored without allocating.

pub fn reserve_exact(&mut self, additional: usize)

Reserve the minimum capacity required to store exactly additional more values.