Module ncollide3d::procedural

Procedural mesh generation.

Re-exports

pub use self::bezier::bezier_curve_at;

pub use self::bezier::bezier_surface_at;

Modules

path

Path generation.

utils

Utilities useful for various generations tasks.

Structs

TriMesh

Geometric description of a mesh.

Enums

IndexBuffer

Different representations of the index buffer.

Functions

bezier_curve

Given a set of control points, generates a (non-rational) Bezier curve.

bezier_surface

Given a set of control points, generates a (non-rational) Bezier surface.

capsule

Generates a capsule.

cone

Generates a cone with a given height and diameter.

cuboid

Generates a cuboid shape with a split index buffer.

cylinder

Generates a cylinder with a given height and diameter.

quad

Adds a double-sided quad to the scene.

quad_with_vertices

Adds a double-sided quad with the specified grid of vertices.

sphere

Generates a UV sphere.

unit_cone

Generates a cone with unit height and diameter.

unit_cuboid

Generates a cuboid shape with a split index buffer.

unit_cylinder

Generates a cylinder with unit height and diameter.

unit_hemisphere

Creates an hemisphere with a diameter of 1.

unit_quad

Adds a double-sided quad with unit size to the scene.

unit_sphere

Generates a UV sphere centered at the origin and with a unit diameter.