Enum ncollide3d::procedural::path::PathSample

pub enum PathSample<N: Real> {
    StartPoint(Point<N>, Vector<N>),
    InnerPoint(Point<N>, Vector<N>),
    EndPoint(Point<N>, Vector<N>),
    EndOfSample,
}

A sample point and its associated tangent.

Variants
StartPoint(Point<N>, Vector<N>)

A point that starts a new path.

InnerPoint(Point<N>, Vector<N>)

A point that is inside of the path currently generated.

EndPoint(Point<N>, Vector<N>)

A point that ends the path currently generated.

EndOfSample

Used when the sampler does not have any other points to generate.