ncollide3d::procedural::path

Trait StrokePattern

pub trait StrokePattern<N: RealField + Copy> {
    // Required method
    fn stroke<C: CurveSampler<N>>(&mut self, sampler: &mut C) -> TriMesh<N>;
}

A pattern that is replicated along a path.

It is responsible of the generation of the whole mesh.

Required Methods

fn stroke<C: CurveSampler<N>>(&mut self, sampler: &mut C) -> TriMesh<N>

Generates the mesh using this pattern and the curve sampled by sampler.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.