Function ncollide3d::procedural::utils::split_index_buffer_and_recover_topology

pub fn split_index_buffer_and_recover_topology<N: RealField>(
    indices: &[Point<u32>], 
    coords: &[Point<N>]
) -> (Vec<Point<Point<u32>>>, Vec<Point<N>>)

Duplicates the indices of each triangle on the given index buffer, giving the same id to each identical vertex.