Module ncollide3d::query::algorithms

Algorithms needed for distance and penetration depth computation.

Re-exports

pub use self::epa3::EPA;

Modules

epa3
Three-dimensional penetration depth queries using the Expanding Polytope Algorithm.

gjk
The Gilbert–Johnson–Keerthi distance algorithm.

special_support_maps
Special support map function like constants and Minkowski sums.

Structs

CSOPoint
A point of a Configuration-Space Obstacle.

VoronoiSimplex
A simplex of dimension up to 3 that uses Voronoï regions for computing point projections.