Module ncollide3d::query::algorithms::gjk

The Gilbert–Johnson–Keerthi distance algorithm.

Enums

GJKResult

Results of the GJK algorithm.

Functions

cast_ray

Casts a ray on a support map using the GJK algorithm.

closest_points

Projects the origin on a shape using the Separating Axis GJK algorithm. The algorithm will stop as soon as the polytope can be proven to be at least max_dist away from the origin.

directional_distance

Compute the normal and the distance that can travel g1 along the direction dir so that g1 and g2 just touch.

eps_tol

The absolute tolerence used by the GJK algorithm.

project_origin

Projects the origin on the boundary of the given shape.