ncollide3d::query::algorithms

Module special_support_maps

Special support map function like constants and Minkowski sums.

Structs

ConstantOrigin

A support mapping that is always equal to the origin, independently from its transform.

DilatedShape

The Minkowski sum of a shape and a ball.