Struct ncollide3d::query::algorithms::CSOPoint

pub struct CSOPoint<N: Real> {
    pub point: Point<N>,
    pub orig1: Point<N>,
    pub orig2: Point<N>,
}

A point of a Configuration-Space Obstacle.

A Configuration-Space Obstacle (CSO) is the result of the Minkowski Difference of two solids. In other words, each of its points correspond to the difference of two point, each belonging to a different solid.

Fields
point: Point<N>

The point on the CSO. This is equal to self.orig1 - self.orig2.

orig1: Point<N>

The original point on the first shape used to compute self.point.

orig2: Point<N>

The original point on the second shape used to compute self.point.

Methods

impl<N: Real> CSOPoint<N>

pub fn new(orig1: Point<N>, orig2: Point<N>) -> Self

Initializes a CSO point with orig1 - orig2.

pub fn new_with_point(point: Point<N>, orig1: Point<N>, orig2: Point<N>) -> Self

Initializes a CSO point with all information provided.

It is assumed, but not checked, that point == orig1 - orig2.

pub fn single_point(point: Point<N>) -> Self

Initializes a CSO point where both original points are equal.

pub fn origin() -> Self

CSO point where all components are set to zero.

pub fn from_shapes_toward<G1: ?Sized, G2: ?Sized>(
m1: &Isometry<N>,
g1: &G1,
m2: &Isometry<N>,
g2: &G2,
dir: &Unit<Vector<N>>
) -> Self where
G1: SupportMap<N>,
G2: SupportMap<N>,

Computes the support point of the CSO of g1 and g2 toward the unit direction dir.

pub fn from_shapes<G1: ?Sized, G2: ?Sized>(
m1: &Isometry<N>,
g1: &G1,
m2: &Isometry<N>,
g2: &G2,
dir: &Vector<N>
) -> Self where
G1: SupportMap<N>,
G2: SupportMap<N>,

Computes the support point of the CSO of g1 and g2 toward the direction dir.

pub fn translate1(&self, dir: &Vector<N>) -> Self

Translate the first original point of this CSO point.

This will apply the same translation to self.point.

pub fn translate2(&self, dir: &Vector<N>) -> Self

Translate the second original point of this CSO point.

This will apply the opposite translation to self.point.