Struct ncollide3d::query::algorithms::VoronoiSimplex

pub struct VoronoiSimplex<N: Real> { /* fields omitted */ }

A simplex of dimension up to 3 that uses Voronoï regions for computing point projections.

Methods

impl<N: Real> VoronoiSimplex<N>

pub fn new() -> VoronoiSimplex<N>

Creates a new empty simplex.

pub fn swap(&mut self, i1: usize, i2: usize)

Swap two vertices of this simplex.

pub fn reset(&mut self, pt: CSOPoint<N>)

Resets this simplex to a single point.

pub fn add_point(&mut self, pt: CSOPoint<N>) -> bool

Add a point to this simplex.

pub fn proj_coord(&self, i: usize) -> N

Retrieves the barycentric coordinate associated to the i-th by the last call to project_origin_and_reduce.

pub fn point(&self, i: usize) -> &CSOPoint<N>

The i-th point of this simplex.

pub fn prev_proj_coord(&self, i: usize) -> N

Retrieves the barycentric coordinate associated to the i-th before the last call to project_origin_and_reduce.

pub fn prev_point(&self, i: usize) -> &CSOPoint<N>

The i-th point of the simplex before the last call to projet_origin_and_reduce.

pub fn project_origin_and_reduce(&mut self) -> Point<N>

Projets the origin on the boundary of this simplex and reduces self the smallest subsimplex containing the origin.

Retruns the result of the projection or Point::origin() if the origin lies inside of the simplex. The state of the samplex before projection is saved, and can be retrieved using the methods prefixed by prev_.

pub fn project_origin(&mut self) -> Point<N>

Compute the projection of the origin on the boundary of this simplex.

pub fn contains_point(&self, pt: &Point<N>) -> bool

Tests if the given point is already a vertex of this simplex.

pub fn dimension(&self) -> usize

The dimension of the smallest subspace that can contain this simplex.

pub fn prev_dimension(&self) -> usize

The dimension of the simplex before the last call to project_origin_and_reduce.

pub fn max_sq_len(&self) -> N

The maximum squared length of the vertices of this simplex.

pub fn modify_pnts(&mut self, f: &Fn(&mut CSOPoint<N>))

Apply a function to all the vertices of this simplex.