Module ncollide3d::query::closest_points_internal

Implementation details of the closest_points function.

Enums

ClosestPoints
Closest points information.

Functions

ball_against_ball
Proximity between balls.

closest_points_internal
Computes the pair of closest points between two shapes.

composite_shape_against_shape
Closest points between a composite shape and any other shape.

line_against_line
Closest points between two segments.

plane_against_support_map
Closest points between a plane and a support-mapped shape (Cuboid, ConvexHull, etc.)

segment_against_segment
Closest points between segments.

segment_against_segment_with_locations
Closest points between two segments.

segment_against_segment_with_locations_nD
Segment-segment closest points computation in an arbitrary dimension.

shape_against_composite_shape
Closest points between a shape and a composite shape.

support_map_against_plane
Closest points between a support-mapped shape (Cuboid, ConvexHull, etc.) and a plane.

support_map_against_support_map
Closest points between support-mapped shapes (Cuboid, ConvexHull, etc.)

support_map_against_support_map_with_params
Closest points between support-mapped shapes (Cuboid, ConvexHull, etc.)