Struct ncollide3d::query::RayIntersectionCostFn

pub struct RayIntersectionCostFn<'a, N: 'a + Real> { /* fields omitted */ }

A search thet selects the objects that has the smallest time of impact with a given ray.

Methods

impl<'a, N: Real> RayIntersectionCostFn<'a, N>

pub fn new(
ray: &'a Ray<N>,
solid: bool,
uvs: bool
) -> RayIntersectionCostFn<'a, N>

Creates a new BestRayInterferenceSearch.