Trait ncollide3d::query::ContactPreprocessor

pub trait ContactPreprocessor<N: RealField + Copy> {
    // Required method
    fn process_contact(
        &self,
        c: &mut Contact<N>,
        kinematic: &mut ContactKinematic<N>,
        is_first: bool,
    ) -> bool;
}

Pre-process a contact before it is added to a contact manifold.

Required Methods

fn process_contact( &self, c: &mut Contact<N>, kinematic: &mut ContactKinematic<N>, is_first: bool, ) -> bool

Process a contact before it is stored into a contact manifold.

Returns false if the contact should be ignored.