Trait ncollide3d::query::RayCast

pub trait RayCast<N: RealField> {
    fn toi_and_normal_with_ray(
        &self, 
        m: &Isometry<N>, 
        ray: &Ray<N>, 
        max_toi: N, 
        solid: bool
    ) -> Option<RayIntersection<N>>;

    fn toi_with_ray(
        &self, 
        m: &Isometry<N>, 
        ray: &Ray<N>, 
        max_toi: N, 
        solid: bool
    ) -> Option<N> { ... }

    fn toi_and_normal_and_uv_with_ray(
        &self, 
        m: &Isometry<N>, 
        ray: &Ray<N>, 
        max_toi: N, 
        solid: bool
    ) -> Option<RayIntersection<N>> { ... }

    fn intersects_ray(&self, m: &Isometry<N>, ray: &Ray<N>, max_toi: N) -> bool { ... }
}

Traits of objects which can be transformed and tested for intersection with a ray.

Required methods

fn toi_and_normal_with_ray(
&self,
m: &Isometry<N>,
ray: &Ray<N>,
max_toi: N,
solid: bool
) -> Option<RayIntersection<N>>

Computes the time of impact, and normal between this transformed shape and a ray.

Loading content...
Provided methods

fn toi_with_ray(
&self,
m: &Isometry<N>,
ray: &Ray<N>,
max_toi: N,
solid: bool
) -> Option<N>

Computes the time of impact between this transform shape and a ray.

fn toi_and_normal_and_uv_with_ray(
&self,
m: &Isometry<N>,
ray: &Ray<N>,
max_toi: N,
solid: bool
) -> Option<RayIntersection<N>>

Computes time of impact, normal, and texture coordinates (uv) between this transformed shape and a ray.

fn intersects_ray(&self, m: &Isometry<N>, ray: &Ray<N>, max_toi: N) -> bool

Tests whether a ray intersects this transformed shape.

Loading content...