Module ncollide3d::query::visitors

Visitors for performing geometric queries exploiting spatial partitioning data structures.

Structs

AABBSetsInterferencesCollector

Spatial partitioning data structure visitor collecting interferences with a given bounding volume.

BoundingVolumeInterferencesCollector

Spatial partitioning data structure visitor collecting interferences with a given bounding volume.

CompositeClosestPointVisitor

Best-first traversal visitor for computing the point closest to a composite shape.

CompositePointContainmentTest

Visitor for checking if a composite shape contains a specific point.

PointInterferencesCollector

Spatial partitioning structure visitor collecting nodes that may contain a given point.

RayInterferencesCollector

Bounding Volume Tree visitor collecting interferences with a given ray.

RayIntersectionCostFnVisitor

Bounding Volume Tree visitor collecting interferences with a given ray.