ncollide3d::shape

Struct Capsule

pub struct Capsule<N> {
    pub half_height: N,
    pub radius: N,
}

SupportMap description of a capsule shape with its principal axis aligned with the y axis.

Fields
half_height: N

The half-height of the capsuleโs cylindrical part.

radius: N

The radius of the capsule.

Implementations

impl<N: RealField + Copy> Capsule<N>

pub fn new(half_height: N, radius: N) -> Capsule<N>

Creates a new capsule.

Arguments:

half_height - the half length of the capsule along the y axis.

radius - radius of the rounded part of the capsule.

pub fn half_height(&self) -> N

๐Deprecated: use the self.half_height public field directly.

The capsule half length along its local y axis.

pub fn height(&self) -> N

The capsule height along its local y axis.

pub fn radius(&self) -> N

๐Deprecated: use the self.radius public field directly.

The radius of the capsuleโs rounded part.

pub fn segment(&self) -> Segment<N>

The segment that, once dilated by self.radius yields this capsule.

pub fn contact_preprocessor(&self) -> impl ContactPreprocessor<N>

The contact preprocessor to be used for contact determination with this capsule.