Struct ncollide3d::shape::Compound

pub struct Compound<N: Real> { /* fields omitted */ }

A compound shape with an aabb bounding volume.

A compound shape is a shape composed of the union of several simpler shape. This is the main way of creating a concave shape from convex parts. Each parts can have its own delta transformation to shift or rotate it with regard to the other shapes.

Methods

impl<N: Real> Compound<N>

pub fn new(shapes: Vec<(Isometry<N>, ShapeHandle<N>)>) -> Compound<N>

Builds a new compound shape.

impl<N: Real> Compound<N>

pub fn shapes(&self) -> &[(Isometry<N>, ShapeHandle<N>)]

The shapes of this compound shape.

pub fn bvt(&self) -> &BVT<usize, AABB<N>>

The optimization structure used by this compound shape.

pub fn bounding_volumes(&self) -> &[AABB<N>]

The shapes bounding volumes.

pub fn aabb_at(&self, i: usize) -> &AABB<N>

The AABB of the i-th shape compositing this compound.