Struct ncollide3d::shape::HeightField

pub struct HeightField<N: RealField> { /* fields omitted */ }

An heightfield implicitly discretized with triangles.

Implementations

impl<N: RealField> HeightField<N>

pub fn new(heights: DMatrix<N>, scale: Vector<N>) -> Self

Initializes a new heightfield with the given heights and a scaling factor.

pub fn nrows(&self) -> usize

The number of rows of this heightfield.

pub fn ncols(&self) -> usize

The number of columns of this heightfield.

pub fn cell_at_point(&self, pt: &Point3<N>) -> Option<(usize, usize)>

The pair of index of the cell containing the vertical projection of the given point.

pub fn x_at(&self, j: usize) -> N

The smallest x coordinate of the j-th column of this heightfield.

pub fn z_at(&self, i: usize) -> N

The smallest z coordinate of the start of the i-th row of this heightfield.

pub fn triangles<'a>(&'a self) -> impl Iterator<Item = Triangle<N>> + 'a

An iterator through all the triangles of this heightfield.

pub fn triangles_at(
&self,
i: usize,
j: usize
) -> (Option<Triangle<N>>, Option<Triangle<N>>)

The two triangles at the cell (i, j) of this heightfield.

Returns None fore triangles that have been removed because of their user-defined status flags (described by the HeightFieldCellStatus bitfield).

pub fn cell_status(&self, i: usize, j: usize) -> HeightFieldCellStatus

The status of the (i, j)-th cell.

pub fn set_cell_status(
&mut self,
i: usize,
j: usize,
status: HeightFieldCellStatus
)

Set the status of the (i, j)-th cell.

pub fn cells_statuses(&self) -> &DMatrix<HeightFieldCellStatus>

The statuses of all the cells of this heightfield.

pub fn cells_statuses_mut(&mut self) -> &mut DMatrix<HeightFieldCellStatus>

The mutable statuses of all the cells of this heightfield.

pub fn heights(&self) -> &DMatrix<N>

The heights of this heightfield.

pub fn scale(&self) -> &Vector<N>

The scale factor applied to this heightfield.

pub fn cell_width(&self) -> N

The width (extent along its local x axis) of each cell of this heightmap, including the scale factor.

pub fn cell_height(&self) -> N

The height (extent along its local z axis) of each cell of this heightmap, including the scale factor.

pub fn unit_cell_width(&self) -> N

The width (extent along its local x axis) of each cell of this heightmap, excluding the scale factor.

pub fn unit_cell_height(&self) -> N

The height (extent along its local z axis) of each cell of this heightmap, excluding the scale factor.

pub fn aabb(&self) -> &AABB<N>

The AABB of this heightmap.

pub fn convert_triangle_feature_id(
&self,
i: usize,
j: usize,
left: bool,
fid: FeatureId
) -> FeatureId

Converts the FeatureID of the left or right triangle at the cell (i, j) into a FeatureId of the whole heightfield.

pub fn map_elements_in_local_aabb(
&self,
aabb: &AABB<N>,
f: &mut impl FnMut(usize, &Triangle<N>, &dyn ContactPreprocessor<N>)
)

Applies the function f to all the triangles of this heightfield intersecting the given AABB.