Struct ncollide3d::shape::TriMesh

pub struct TriMesh<N: RealField> { /* fields omitted */ }

A 3d triangle mesh.

Implementations

impl<N: RealField> TriMesh<N>

pub fn new(
points: Vec<Point<N>>,
indices: Vec<Point3<usize>>,
uvs: Option<Vec<Point2<N>>>
) -> TriMesh<N>

Builds a new mesh.

pub fn aabb(&self) -> &AABB<N>

The triangle mesh's AABB.

pub fn points(&self) -> &[Point<N>]

The points of this mesh.

pub fn faces(&self) -> &[TriMeshFace<N>]

The faces of this mesh.

pub fn edges(&self) -> &[TriMeshEdge]

The edges of this mesh.

pub fn vertices(&self) -> &[TriMeshVertex]

The vertices of this mesh.

pub fn transform_by(&mut self, transform: &Isometry<N>)

Applies in-place a transformation to this triangle mesh.

pub fn transformed(self, transform: &Isometry<N>) -> Self

Applies a transformation to this triangle mesh.

pub fn scale_by(&mut self, scale: &Vector<N>)

Applies in-place a non-uniform scale to this triangle mesh.

pub fn scaled(self, scale: &Vector<N>) -> Self

Applies a non-uniform scale to this triangle mesh.

pub fn oriented(&self) -> bool

Whether this trimesh is considered is oriented or not.

By default a trimesh is not oriented.

pub fn set_oriented(&mut self, oriented: bool)

Whether this trimesh is considered as oriented or not.

This is determined at the initialization of the trimesh.

pub fn face_containing_feature(&self, id: FeatureId) -> usize

Face containing feature.

pub fn edge_segment(&self, i: usize) -> Segment<N>

The segment of the i-th edge on this triangle mesh.

pub fn uvs(&self) -> Option<&[Point2<N>]>

The texture coordinates of this mesh.

pub fn adj_vertex_list(&self) -> &[usize]

The adjacent vertices list of this mesh.

Use TriMeshVertex.adj_vertices to index this. Elements are indexes into the vertices list.

pub fn adj_face_list(&self) -> &[usize]

The adjacent vertices list of this mesh.

Use TriMeshVertex.adj_faces to index this. Elements are indexes into the faces list.

pub fn triangle_at(&self, i: usize) -> Triangle<N>

Gets the i-th mesh element.

pub fn is_backface(&self, feature: FeatureId) -> bool

Returns true if the given feature is a FeatureId::Face and identifies a backface of this trimesh.

pub fn bvt(&self) -> &BVT<usize, AABB<N>>

The optimization structure used by this triangle mesh.

pub fn vertex_tangent_cone_contains_dir(
&self,
i: usize,
deformations: Option<&[N]>,
dir: &Unit<Vector<N>>
) -> bool

Tests that the given dir is on the tangent cone of the ith vertex of this mesh.

pub fn vertex_tangent_cone_polar_contains_dir(
&self,
i: usize,
dir: &Unit<Vector<N>>,
sin_ang_tol: N
) -> bool

Tests that the given dir is on the polar of the tangent cone of the ith vertex of this mesh.

pub fn edge_tangent_cone_contains_dir(
&self,
i: usize,
deformations: Option<&[N]>,
dir: &Unit<Vector<N>>
) -> bool

Tests that the given dir is on the tangent cone of the ith edge of this mesh.

pub fn edge_tangent_cone_polar_contains_orthogonal_dir(
&self,
i: usize,
dir: &Unit<Vector<N>>,
sin_ang_tol: N
) -> bool

Tests that the given dir is on the polar of the tangent cone of the ith edge of this mesh.

The dir is assumed to be orthogonal to the edge.

pub fn edge_tangent_cone_polar_contains_dir(
&self,
i: usize,
dir: &Unit<Vector<N>>,
sin_ang_tol: N,
_cos_ang_tol: N
) -> bool

Tests that the given dir is on the polar of the tangent cone of the ith edge of this mesh.

pub fn face_tangent_cone_contains_dir(
&self,
i: usize,
deformations: Option<&[N]>,
dir: &Unit<Vector<N>>
) -> bool

Tests that the given dir is on the tangent cone of the ith face of this mesh.

pub fn face_tangent_cone_polar_contains_dir(
&self,
i: usize,
dir: &Unit<Vector<N>>,
cos_ang_tol: N
) -> bool

Checks if the polar of the tangent cone of the i-th face of this triangle mesh contains the specified direction within an angular tolerence.

pub fn tangent_cone_polar_contains_dir(
&self,
feature: FeatureId,
dir: &Unit<Vector<N>>,
sin_ang_tol: N,
cos_ang_tol: N
) -> bool

Checks if the polar of the tangent cone of the specified feature of this triangle mesh contains the specified direction within an angular tolerence.