Trait ncollide3d::utils::IsometryOps

pub trait IsometryOps<N: RealField + Copy> {
    // Required methods
    fn absolute_transform_vector(&self, v: &Vector<N>) -> Vector<N>;
    fn inverse_transform_unit_vector(
        &self,
        v: &Unit<Vector<N>>
    ) -> Unit<Vector<N>>;
    fn lerp_slerp(&self, other: &Self, t: N) -> Self;
}

Extra operations with isometries.

Required Methods

fn absolute_transform_vector(&self, v: &Vector<N>) -> Vector<N>

Transform a vector by the absolute value of the homogeneous matrix equivalent to self.

fn inverse_transform_unit_vector(&self, v: &Unit<Vector<N>>) -> Unit<Vector<N>>

Transform a unit vector by the inverse of self.

fn lerp_slerp(&self, other: &Self, t: N) -> Self

Interpolates between two isometries, using LERP for the translation part and SLERP for the rotation part.

Object Safety

This trait is not object safe.