Enum ncollide3d::world::GeometricQueryType

pub enum GeometricQueryType<N: Real> {
    Contacts(N, N),
    Proximity(N),
}

The kind of query a CollisionObject may be involved on.

The following queries are executed for a given pair of GeometricQueryType associated with two collision objects:

Contacts + Contacts = exact contact point coputation.

Contacts + Proximity = proximity test only.

Proximity + Proximity = proximity test only.

Variants
Contacts(N, N)

This objects can respond to both contact point computation and proximity queries.

Proximity(N)

This object can respond to proximity tests only.

Methods

impl<N: Real> GeometricQueryType<N>

pub fn query_limit(&self) -> N

The numerical distance limit of relevance for this query.

If two objects are separated by a distance greater than the sum of their respective query_limit, the corresponding query will not by performed. For proximity queries, non-intersecting object closer than a distance equal to the sum of their query_limit will be reported as Proximity::WithinMargin.

pub fn contact_queries_to_prediction(
self,
other: Self
) -> Option<ContactPrediction<N>>

Given two contact query types, returns the corresponding contact prediction parameters.

Returns None if any of self or other is not a GeometricQueryType::Contacts.

pub fn is_contacts_query(&self) -> bool

Returns true if this is a contacts query type.

pub fn is_proximity_query(&self) -> bool

Returns true if this is a proximity query type.