Struct ncollide3d::world::CollisionObject

pub struct CollisionObject<N: Real, T> { /* fields omitted */ }

A stand-alone object that has a position and a shape.

Methods

impl<N: Real, T> CollisionObject<N, T>

pub fn new(
handle: CollisionObjectHandle,
proxy_handle: ProxyHandle,
position: Isometry<N>,
shape: ShapeHandle<N>,
groups: CollisionGroups,
query_type: GeometricQueryType<N>,
data: T
) -> CollisionObject<N, T>

Creates a new collision object.

pub fn handle(&self) -> CollisionObjectHandle

The collision object unique handle.

pub fn proxy_handle(&self) -> ProxyHandle

The collision object's broad phase proxy unique identifier.

pub fn position(&self) -> &Isometry<N>

The collision object position.

pub fn set_position(&mut self, pos: Isometry<N>)

Sets the position of the collision object.

pub fn shape(&self) -> &ShapeHandle<N>

The collision object shape.

pub fn collision_groups(&self) -> &CollisionGroups

The collision groups of the collision object.

pub fn query_type(&self) -> GeometricQueryType<N>

The kind of queries this collision object is expected to .

pub fn data(&self) -> &T

Reference to the user-defined data associated to this object.

pub fn data_mut(&mut self) -> &mut T

Mutable reference to the user-defined data associated to this object.