Struct ncollide3d::world::CollisionWorld

pub struct CollisionWorld<N: Real, T> { /* fields omitted */ }

A world that handles collision objects.

Methods

impl<N: Real, T> CollisionWorld<N, T>

pub fn new(margin: N) -> CollisionWorld<N, T>

Creates a new collision world.

pub fn add(
&mut self,
position: Isometry<N>,
shape: ShapeHandle<N>,
collision_groups: CollisionGroups,
query_type: GeometricQueryType<N>,
data: T
) -> CollisionObjectHandle

Adds a collision object to the world.

pub fn update(&mut self)

Updates the collision world.

This executes the whole collision detection pipeline:

Clears the event pools.

Executes the broad phase first.

Executes the narrow phase.

pub fn clear_events(&mut self)

Empty the contact and proximity event pools.

pub fn remove(&mut self, handles: &[CollisionObjectHandle])

Removed the specified set of collision objects from the world.

Panics of any handle is invalid, or if the list contains duplicates.

pub fn set_position(&mut self, handle: CollisionObjectHandle, pos: Isometry<N>)

Sets the position the collision object attached to the specified object.

pub fn register_broad_phase_pair_filter<F>(&mut self, name: &str, filter: F) where
F: BroadPhasePairFilter<N, T>,

Adds a filter that tells if a potential collision pair should be ignored or not.

The proximity filter returns false for a given pair of collision objects if they should be ignored by the narrow phase. Keep in mind that modifying the proximity filter will have a non-trivial overhead during the next update as it will force re-detection of all collision pairs.

pub fn unregister_broad_phase_pair_filter(&mut self, name: &str)

Removes the pair filter named name.

pub fn perform_broad_phase(&mut self)

Executes the broad phase of the collision detection pipeline.

pub fn perform_narrow_phase(&mut self)

Executes the narrow phase of the collision detection pipeline.

pub fn set_narrow_phase(
&mut self,
narrow_phase: Box<NarrowPhase<N, T>>
) -> Box<NarrowPhase<N, T>>

Sets a new narrow phase and returns the previous one.

Keep in mind that modifying the narrow-pase will have a non-trivial overhead during the next update as it will force re-detection of all collision pairs and their associated contacts.

Important traits for ContactPairs<'a, N, T>
impl<'a, N: Real, T> Iterator for ContactPairs<'a, N, T> type Item = (&'a CollisionObject<N, T>, &'a CollisionObject<N, T>, &'a ContactAlgorithm<N>);

pub fn contact_pairs(&self) -> ContactPairs<N, T>

Iterates through all the contact pairs detected since the last update.

Important traits for ProximityPairs<'a, N, T>
impl<'a, N: Real, T> Iterator for ProximityPairs<'a, N, T> type Item = (&'a CollisionObject<N, T>, &'a CollisionObject<N, T>, &'a ProximityAlgorithm<N>);

pub fn proximity_pairs(&self) -> ProximityPairs<N, T>

Iterates through all the proximity pairs detected since the last update.

Important traits for ContactManifolds<'a, N, T>
impl<'a, N: Real, T> Iterator for ContactManifolds<'a, N, T> type Item = (&'a CollisionObject<N, T>, &'a CollisionObject<N, T>, &'a ContactManifold<N>);

pub fn contact_manifolds(&self) -> ContactManifolds<N, T>

Iterates through every contact detected since the last update.

Important traits for CollisionObjects<'a, N, T>
impl<'a, N: 'a + Real, T: 'a> Iterator for CollisionObjects<'a, N, T> type Item = &'a CollisionObject<N, T>;

pub fn collision_objects(&self) -> CollisionObjects<N, T>

Iterates through all collision objects.

pub fn collision_object(
&self,
handle: CollisionObjectHandle
) -> Option<&CollisionObject<N, T>>

Returns a reference to the collision object identified by its handle.

pub fn collision_object_mut(
&mut self,
handle: CollisionObjectHandle
) -> Option<&mut CollisionObject<N, T>>

Returns a mutable reference to the collision object identified by its handle.

pub fn set_collision_groups(
&mut self,
handle: CollisionObjectHandle,
groups: CollisionGroups
)

Sets the collision groups of the given collision object.

Important traits for InterferencesWithRay<'a, 'b, N, T>
impl<'a, 'b, N: Real, T> Iterator for InterferencesWithRay<'a, 'b, N, T> type Item = (&'a CollisionObject<N, T>, RayIntersection<N>);

pub fn interferences_with_ray<'a, 'b>(
&'a self,
ray: &'b Ray<N>,
groups: &'b CollisionGroups
) -> InterferencesWithRay<'a, 'b, N, T>

Computes the interferences between every rigid bodies on this world and a ray.

Important traits for InterferencesWithPoint<'a, 'b, N, T>
impl<'a, 'b, N: Real, T> Iterator for InterferencesWithPoint<'a, 'b, N, T> type Item = &'a CollisionObject<N, T>;

pub fn interferences_with_point<'a, 'b>(
&'a self,
point: &'b Point<N>,
groups: &'b CollisionGroups
) -> InterferencesWithPoint<'a, 'b, N, T>

Computes the interferences between every rigid bodies of a given broad phase, and a point.

Important traits for InterferencesWithAABB<'a, 'b, N, T>
impl<'a, 'b, N: Real, T> Iterator for InterferencesWithAABB<'a, 'b, N, T> type Item = &'a CollisionObject<N, T>;

pub fn interferences_with_aabb<'a, 'b>(
&'a self,
aabb: &'b AABB<N>,
groups: &'b CollisionGroups
) -> InterferencesWithAABB<'a, 'b, N, T>

Computes the interferences between every rigid bodies of a given broad phase, and a aabb.

pub fn contact_events(&self) -> &ContactEvents

The contact events pool.

pub fn proximity_events(&self) -> &ProximityEvents

The proximity events pool.