Crate ncollide3d

ncollide

ncollide is a 2 and 3-dimensional collision detection library written with the rust programming language.

As its name suggests, it is generic wrt the dimension: it works with both 2-dimensional and 3-dimensional shapes.

The official user guide is available here. The rustdoc documentation is available here.

Compilation

You will need the last stable build of the rust compiler and the official package manager: cargo.

Simply add the following to your Cargo.toml file:

[dependencies]
ncollide2d = "0.23" # For 2D collision detection.
ncollide3d = "0.23" # For 3D collision detection.

Features

dynamic bounding volume tree based broad phase

ball vs. ball collision detection,

plane vs. any convex object collision detection.

collision detection between arbitrary convex objects

compound shapes

ray-casting

time of impact computation for objects without rotational movement (compound vs. compound is not yet implemented)

And various traits for collision detectors and broad phase collision detection.

Re-exports

pub extern crate nalgebra as na;

pub extern crate simba;

pub use crate::pipeline::broad_phase;

pub use crate::pipeline::narrow_phase;

pub use crate::pipeline::world;

Modules

bounding_volume
Bounding volumes.

interpolation
Interpolation of the motion of an object.

math
Compilation flags dependent aliases for mathematical types.

partitioning
Spatial partitioning tools.

pipeline
Persistent and time-coherent collision detection.

procedural
Procedural mesh generation.

query
Non-persistent geometric queries.

shape
Collision shapes supported by ncollide.

transformation
Transformation, simplification and decomposition of meshes.

utils
Various unsorted geometrical and logical operators.