Module ncollide3d::bounding_volume

Bounding volumes.

Structs

AABB
An Axis Aligned Bounding Box.

BoundingSphere
A Bounding Sphere.

SpatializedNormalCone
The combination of an AABB with a circular cone to bound both the space occupied by an geometry and its normals.

Enums

CircularCone
A cone with a circular basis and its apex at the origin.

Traits

BoundingVolume
Trait of bounding volumes.

HasBoundingVolume
Traits of objects having a bounding volume.

Functions

aabb
Computes the axis-aligned bounding box of a shape g transformed by m.

ball_aabb
Computes the Axis-Aligned Bounding Box of a ball transformed by center.

bounding_sphere
Computes the bounding sphere of a shape g transformed by m.

local_aabb
Computes the axis-aligned bounding box of a shape g.

local_bounding_sphere
Computes the bounding sphere of a shape g.

local_point_cloud_aabb
Computes the AABB of a set of points.

local_support_map_aabb
Computes the AABB of an support mapped shape.

point_cloud_aabb
Computes the AABB of a set of points transformed by m.

point_cloud_bounding_sphere
Computes a bounding sphere of the specified set of point.

point_cloud_bounding_sphere_with_center
Computes the bounding sphere of a set of point, given its center.

support_map_aabb
Computes the AABB of an support mapped shape.