Struct ncollide3d::bounding_volume::AABB

pub struct AABB<N: RealField> {
    pub mins: Point<N>,
    pub maxs: Point<N>,
}

An Axis Aligned Bounding Box.

Fields
mins: Point<N>maxs: Point<N>
Implementations

impl<N: RealField> AABB<N>

pub fn new(mins: Point<N>, maxs: Point<N>) -> AABB<N>

Creates a new AABB.

Arguments:

mins - position of the point with the smallest coordinates.

maxs - position of the point with the highest coordinates. Each component of mins must be smaller than the related components of maxs.

pub fn new_invalid() -> Self

Creates an invalid AABB with mins components set to N::max_values and maxscomponents set to -N::max_values.

This is often used as the initial values of some AABB merging algorithms.

pub fn from_half_extents(center: Point<N>, half_extents: Vector<N>) -> Self

Creates a new AABB from its center and its half-extents.

pub fn from_points<'a, I>(pts: I) -> Self where
I: IntoIterator<Item = &'a Point<N>>,

Creates a new AABB from a set of points.

pub fn mins(&self) -> &Point<N>

👎 Deprecated:
use the .mins public field instead.

Reference to the AABB point with the smallest components along each axis.

pub fn maxs(&self) -> &Point<N>

👎 Deprecated:
use the .maxs public field instead.

Reference to the AABB point with the biggest components along each axis.

pub fn center(&self) -> Point<N>

The center of this AABB.

pub fn half_extents(&self) -> Vector<N>

The half extents of this AABB.

pub fn extents(&self) -> Vector<N>

The extents of this AABB.

pub fn take_point(&mut self, pt: Point<N>)

Enlarges this AABB so it also contains the point pt.

pub fn transform_by(&self, m: &Isometry<N>) -> Self

Computes the AABB bounding self transformed by m.

pub fn bounding_sphere(&self) -> BoundingSphere<N>

The smallest bounding sphere containing this AABB.

pub fn contains_local_point(&self, point: &Point<N>) -> bool

impl<N: RealField> AABB<N>

pub fn clip_line_parameters(
&self,
orig: &Point<N>,
dir: &Vector<N>
) -> Option<(N, N)>

Computes the parameters of the two intersection points between a line and this AABB.

The parameters are such that the point are given by orig + dir * parameter. Returns None if there is no intersection.

pub fn clip_line(&self, orig: &Point<N>, dir: &Vector<N>) -> Option<Segment<N>>

Computes the intersection segment between a line and this AABB.

Returns None if there is no intersection.

pub fn clip_ray_parameters(&self, ray: &Ray<N>) -> Option<(N, N)>

Computes the parameters of the two intersection points between a ray and this AABB.

The parameters are such that the point are given by ray.orig + ray.dir * parameter. Returns None if there is no intersection.

pub fn clip_ray(&self, ray: &Ray<N>) -> Option<Segment<N>>

Computes the intersection segment between a ray and this AABB.

Returns None if there is no intersection.