Module ncollide3d::partitioning

Spatial partitioning tools.

Structs

BVT
A Bounding Volume Tree.

DBVT
A bounding volume hierarchy on which objects can be added or removed after construction.

DBVTLeaf
Leaf of a Dynamic Bounding Volume Tree.

DBVTLeafId
The unique identifier of a DBVT leaf.

Enums

BVHImpl
An enum grouping references to all the BVH implementations on ncollide.

BVTNodeId
The identifier of a BVT node.

BestFirstVisitStatus
The next action to be taken by a BVH traversal algorithm after having visited a node with some data.

BinaryPartition
Result of a binary partition.

DBVTNodeId
The identifier of a node of the DBVT.

VisitStatus
The status of the spatial partitioning structure traversal.

Traits

BVH
Trait implemented by Bounding Volume Hierarchy.

BestFirstVisitor
Trait implemented by cost functions used by the best-first search on a BVT.

SimultaneousVisitor
Trait implemented by visitor called during a simultaneous spatial partitioning data structure tarversal.

Visitor
Trait implemented by visitor called during the traversal of a spatial partitioning data structure.