ncollide3d::partitioning

Struct BVT

pub struct BVT<T, BV> { /* private fields */ }

A Bounding Volume Tree.

Implementations

impl<T, BV> BVT<T, BV>

pub fn new_with_partitioning<F: FnMut(usize, Vec<(T, BV)>) -> (BV, BinaryPartition<T, BV>)>( elements: Vec<(T, BV)>, partitioning: &mut F, ) -> BVT<T, BV>

๐Deprecated: please use from_partitioning instead

Builds a bounding volume tree using the specified partitioning function.

pub fn from_partitioning( elements: Vec<(T, BV)>, partitioning: &mut impl FnMut(usize, Vec<(T, BV)>) -> (BV, BinaryPartition<T, BV>), ) -> BVT<T, BV>

Builds a bounding volume tree using the specified partitioning function.

pub fn leaves(&self) -> &[BVTLeaf<T, BV>]

The set of leaves on this BVT.

pub fn leaf(&self, i: usize) -> &BVTLeaf<T, BV>

Referenceto the i-th leaf of this BVT.

pub fn root_bounding_volume(&self) -> Option<&BV>

Reference to the bounding volume of the tree root.

pub fn set_leaf_bounding_volume<N: RealField + Copy>( &mut self, i: usize, bv: BV, refit_now: bool, )
where BV: BoundingVolume<N>,

Set the bounding volume of the i-th leaf.

If refit_now is true, the bounding volumes of all the ancestors of the modifiad leaf will be updated as well to enclose the new leaf bounding volume. If refit_now is false, no ancestor update will be performed until the .refit() method is called. This is useful to refit the tree only once after several leaf bounding volume modifications.

pub fn refit<N: RealField + Copy>(&mut self, margin: N)
where BV: BoundingVolume<N>,

Refits the bounding volumes so that all node of the BVT have boundin volumes that enclose their children.

This must be called to ensure the BVT is in a valid state after several calls to .set_leaf_bounding_volume(_, _, false). Every bounding volume created during this update will be enlarged by a margin of margin. The larger this margin here, the looser will the resulting AABB will be, but the less frequent future updates will be necessary. Setting a margin equal to 0.0 is allowed.

impl<T, BV> BVT<T, BV>

pub fn new_balanced<N>(leaves: Vec<(T, BV)>) -> BVT<T, BV>
where N: RealField + Copy, BV: BoundingVolume<N> + Clone,

Creates a balanced BVT.

pub fn median_partitioning_with_centers<N, F: FnMut(&T, &BV) -> Point<N>>( depth: usize, leaves: Vec<(T, BV)>, center: &mut F, ) -> (BV, BinaryPartition<T, BV>)
where N: RealField + Copy, BV: BoundingVolume<N> + Clone,

Construction function for a kdree to be used with BVT::from_partitioning.

pub fn median_partitioning<N>( depth: usize, leaves: Vec<(T, BV)>, ) -> (BV, BinaryPartition<T, BV>)
where N: RealField + Copy, BV: BoundingVolume<N> + Clone,

Construction function for a kdree to be used with BVT::from_partitioning.