ncollide3d::partitioning

Struct DBVT

pub struct DBVT<N: RealField + Copy, T, BV> { /* private fields */ }

A bounding volume hierarchy on which objects can be added or removed after construction.

Implementations

impl<N: RealField + Copy, T, BV: BoundingVolume<N>> DBVT<N, T, BV>

pub fn new() -> DBVT<N, T, BV>

Creates a new empty dynamic bonding volume hierarchy.

pub fn root_bounding_volume(&self) -> Option<&BV>

The bounding volume of the root of this DBVT.

Returns None if the DBVT is empty.

pub fn is_empty(&self) -> bool

Indicates whether this DBVT empty.

pub fn insert(&mut self, leaf: DBVTLeaf<N, T, BV>) -> DBVTLeafId

Inserts a leaf into this DBVT.

pub fn remove(&mut self, leaf_id: DBVTLeafId) -> DBVTLeaf<N, T, BV>

Removes a leaf from this DBVT.

Panics if the provided leaf is not attached to this DBVT.

pub fn get(&self, DBVTLeafId: DBVTLeafId) -> Option<&DBVTLeaf<N, T, BV>>

Gets the given leaf if it exists.