ncollide3d::partitioning

Trait BVH

pub trait BVH<T, BV> {
    type Node: Copy;

    // Required methods
    fn root(&self) -> Option<Self::Node>;
    fn num_children(&self, node: Self::Node) -> usize;
    fn child(&self, i: usize, node: Self::Node) -> Self::Node;
    fn content(&self, node: Self::Node) -> (&BV, Option<&T>);

    // Provided methods
    fn visit(&self, visitor: &mut impl Visitor<T, BV>) { ... }
    fn visit_bvtt(
        &self,
        other: &impl BVH<T, BV>,
        visitor: &mut impl SimultaneousVisitor<T, BV>,
    ) { ... }
    fn best_first_search<N, BFS>(
        &self,
        visitor: &mut BFS,
    ) -> Option<(Self::Node, BFS::Result)>
       where N: RealField + Copy,
             BFS: BestFirstVisitor<N, T, BV> { ... }
}

Trait implemented by Bounding Volume Hierarchy.

Required Associated Types

type Node: Copy

Type of a node identifiers on this BVH.

Required Methods

fn root(&self) -> Option<Self::Node>

The root of the BVH.

fn num_children(&self, node: Self::Node) -> usize

The number of children of the given node.

fn child(&self, i: usize, node: Self::Node) -> Self::Node

The i-th child of the given node.

fn content(&self, node: Self::Node) -> (&BV, Option<&T>)

The bounding volume and data contained by the given node.

Provided Methods

fn visit(&self, visitor: &mut impl Visitor<T, BV>)

Traverses this BVH using a visitor.

fn visit_bvtt( &self, other: &impl BVH<T, BV>, visitor: &mut impl SimultaneousVisitor<T, BV>, )

Visits the bounding volume test tree implicitly formed with other.

fn best_first_search<N, BFS>( &self, visitor: &mut BFS, ) -> Option<(Self::Node, BFS::Result)>
where N: RealField + Copy, BFS: BestFirstVisitor<N, T, BV>,

Performs a best-first-search on the BVH.

Returns the content of the leaf with the smallest associated cost, and a result of user-defined type.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.