ncollide3d::partitioning

Trait Visitor

pub trait Visitor<T, BV> {
    // Required method
    fn visit(&mut self, bv: &BV, data: Option<&T>) -> VisitStatus;
}

Trait implemented by visitor called during the traversal of a spatial partitioning data structure.

Required Methods

fn visit(&mut self, bv: &BV, data: Option<&T>) -> VisitStatus

Execute an operation on the content of a node of the spatial partitioning structure.

Returns whether the traversal should continue on the node’s children, if it should not continue on those children, or if the whole traversal should be exited early.