Module ncollide3d::pipeline::glue

Glue code between each part of the collision-detection pipeline.

Structs

FirstInterferenceWithRay
Return structure for first_interference_with_ray

InterferencesWithAABB
Iterator through all the objects on the world which bounding volume intersects a specific AABB.

InterferencesWithPoint
Iterator through all the objects on the world that intersect a specific point.

InterferencesWithRay
Iterator through all the objects on the world that intersect a specific ray.

Functions

create_proxies
Registers a collision object handle so it can be taken into acconut by the broad-phase and the narrow-phase.

default_broad_phase
Allocate a default broad-phase, configured with a default coherence margin (set to 0.01).

default_interaction_graph
Allocate a default interaction graph.

default_narrow_phase
Allocate a default narrow-phase, configured with the default contact and proximity dispatchers.

first_interference_with_ray
Returns an the closest collision object intersecting with the given ray.

interferences_with_aabb
Returns an iterator yielding all the collision objects with an AABB intersecting with the given AABB.

interferences_with_point
Returns an iterator yielding all the collision objects containing the given point.

interferences_with_ray
Returns an iterator yielding all the collision objects intersecting with the given ray.

perform_all_pipeline
Performs the broad-phase and the narrow-phase.

perform_broad_phase
Performs the broad-phase.

perform_narrow_phase
Performs the narrow-phase.

remove_proxies
Free all the resources allocated by the broad-phase and the interaction graph for the given proxy handles.