ncollide3d::pipeline::glue

Function create_proxies

pub fn create_proxies<'a, N: RealField + Copy, Handle: CollisionObjectHandle>(
    handle: Handle,
    broad_phase: &mut (impl BroadPhase<N, AABB<N>, Handle> + ?Sized),
    interactions: &mut InteractionGraph<N, Handle>,
    position: &Isometry<N>,
    shape: &(impl Shape<N> + ?Sized),
    query_type: GeometricQueryType<N>,
) -> (BroadPhaseProxyHandle, CollisionObjectGraphIndex)

Registers a collision object handle so it can be taken into acconut by the broad-phase and the narrow-phase.

This tells the broad-phase and the interaction graph exists and allow them to perform some setup to accomodate for a new collision object with the given handle, and known to currently have the given position, shape and query_type. The result of this registration is a pair of handles that must be stored by the user as the will be needed for various queries (on the broad-phase and interaction graph), as well as for freeing (using remove_proxies) the resources allocated here.