Module ncollide3d::pipeline::narrow_phase

Persistent collision detection algorithms to compute contact points.

Structs

BallBallManifoldGenerator
Collision detector between two balls.

BallBallProximityDetector
Proximity detector between two balls.

BallConvexPolyhedronManifoldGenerator
Collision detector between two balls.

CapsuleCapsuleManifoldGenerator
Collision detector between a concave shape and another shape.

CapsuleShapeManifoldGenerator
Collision detector between a concave shape and another shape.

CompositeShapeCompositeShapeManifoldGenerator
Collision detector between a concave shape and another shape.

CompositeShapeShapeManifoldGenerator
Collision detector between a concave shape and another shape.

CompositeShapeShapeProximityDetector
Proximity detector between a concave shape and another shape.

ConvexPolyhedronConvexPolyhedronManifoldGenerator

DefaultContactDispatcher
Collision dispatcher for shapes defined by ncollide_entities.

DefaultProximityDispatcher
Proximity dispatcher for shapes defined by ncollide_entities.

EventPool
A set of events.

HeightFieldShapeManifoldGenerator
Collision detector between an heightfield and another shape.

InteractionGraph
A graph where nodes are collision objects and edges are contact or proximity algorithms.

NarrowPhase
Collision detector dispatcher for collision objects.

PlaneBallManifoldGenerator
Collision detector between g1 plane and g1 shape implementing the SupportMap trait.

PlaneConvexPolyhedronManifoldGenerator
Collision detector between g1 plane and g1 shape implementing the SupportMap trait.

PlaneSupportMapProximityDetector
Proximity detector between a plane and a shape implementing the SupportMap trait.

ProximityEvent
Events occuring when two collision objects start or stop being in close proximity, contact, or disjoint.

SupportMapPlaneProximityDetector
Proximity detector between a plane and a shape implementing the SupportMap trait.

SupportMapSupportMapProximityDetector
Persistent proximity detector between two shapes having a support mapping function.

TriMeshTriMeshManifoldGenerator
Collision detector between a concave shape and another shape.

Enums

ContactEvent
Events occuring when two collision objects start or stop being in contact (or penetration).

Interaction
An interaction between two collision objects.

Traits

ContactDispatcher

ContactManifoldGenerator
An algorithm to compute contact points, normals and penetration depths between two specific objects.

ProximityDetector
Trait implemented by algorithms that determine if two objects are in close proximity.

ProximityDispatcher

Type Definitions

CollisionObjectGraphIndex
Index of a node of the interaction graph.

ContactAlgorithm

ContactEvents
A set of contact events.

ProximityAlgorithm

ProximityEvents
A set of proximity events.

TemporaryInteractionIndex
Temporary index to and edge of the interaction graph.