ncollide3d::pipeline::narrow_phase

Trait ContactManifoldGenerator

pub trait ContactManifoldGenerator<N: RealField + Copy>:
    Any
    + Send
    + Sync {
    // Required method
    fn generate_contacts(
        &mut self,
        dispatcher: &dyn ContactDispatcher<N>,
        ma: &Isometry<N>,
        a: &dyn Shape<N>,
        proc1: Option<&dyn ContactPreprocessor<N>>,
        mb: &Isometry<N>,
        b: &dyn Shape<N>,
        proc2: Option<&dyn ContactPreprocessor<N>>,
        prediction: &ContactPrediction<N>,
        manifold: &mut ContactManifold<N>,
    ) -> bool;

    // Provided method
    fn init_manifold(&self) -> ContactManifold<N> { ... }
}

An algorithm to compute contact points, normals and penetration depths between two specific objects.

Required Methods

fn generate_contacts( &mut self, dispatcher: &dyn ContactDispatcher<N>, ma: &Isometry<N>, a: &dyn Shape<N>, proc1: Option<&dyn ContactPreprocessor<N>>, mb: &Isometry<N>, b: &dyn Shape<N>, proc2: Option<&dyn ContactPreprocessor<N>>, prediction: &ContactPrediction<N>, manifold: &mut ContactManifold<N>, ) -> bool

Runs the collision detection on two objects. It is assumed that the same collision detector (the same structure) is always used with the same pair of objects.

Returns false if persisting this algorithm for re-use is unlikely to improve performance, e.g. due to the objects being distant. Note that if the ContactManifoldGenerator would likely be immediately reconstructed in the next time-step, dropping it is sub-optimal regardless.

Provided Methods

fn init_manifold(&self) -> ContactManifold<N>

Generate an empty contact manifold configured as required by this contact manifold generator.