Trait ncollide3d::pipeline::narrow_phase::ProximityDetector

pub trait ProximityDetector<N: RealField + Copy>: Any + Send + Sync {
    fn update(
        &mut self,
        dispatcher: &dyn ProximityDispatcher<N>,
        ma: &Isometry<N>,
        a: &dyn Shape<N>,
        mb: &Isometry<N>,
        b: &dyn Shape<N>,
        margin: N
    ) -> Option<Proximity>;
}

Trait implemented by algorithms that determine if two objects are in close proximity.

Required Methods

fn update(
&mut self,
dispatcher: &dyn ProximityDispatcher<N>,
ma: &Isometry<N>,
a: &dyn Shape<N>,
mb: &Isometry<N>,
b: &dyn Shape<N>,
margin: N
) -> Option<Proximity>

Runs the proximity detection on two objects. It is assumed that the same proximity detector (the same instance) is always used with the same pair of object.