ncollide3d::pipeline

Module object

Definition of collision objects and some of their properties.

Structs

CollisionGroups

Groups of collision used to filter which object interact with which other one.

CollisionGroupsPairFilter

A collision filter based collision groups.

CollisionObject

A stand-alone object that has a position and a shape.

CollisionObjectSlab

A set of collision objects that can be indexed by collision object handles.

CollisionObjectSlabHandle

The unique identifier of a collision object stored in a CollisionObjectSlab structure.

CollisionObjectUpdateFlags

Flags indicating what changed in a collision object since the last collision world update.

CollisionObjects

An iterator yielding references to collision objects.

Enums

GeometricQueryType

The kind of query a CollisionObject may be involved on.

Traits

CollisionObjectHandle

Trait implemented by a handle indentifying a collision object.

CollisionObjectRef

Trait implemented by collision objects.

CollisionObjectSet

Trait implemented by sets of collision objects.