ncollide3d::pipeline::object

Struct CollisionObject

pub struct CollisionObject<N: RealField + Copy, T> { /* private fields */ }

A stand-alone object that has a position and a shape.

Implementations

impl<N: RealField + Copy, T> CollisionObject<N, T>

pub fn new( proxy_handle: Option<BroadPhaseProxyHandle>, graph_index: Option<CollisionObjectGraphIndex>, position: Isometry<N>, shape: ShapeHandle<N>, groups: CollisionGroups, query_type: GeometricQueryType<N>, data: T, ) -> CollisionObject<N, T>

Creates a new collision object.

pub fn graph_index(&self) -> Option<CollisionObjectGraphIndex>

The collision object non-stable graph index.

This index may change whenever a collision object is removed from the world.

pub fn set_graph_index(&mut self, index: Option<CollisionObjectGraphIndex>)

Sets the collision object unique but non-stable graph index.

pub fn update_flags_mut(&mut self) -> &mut CollisionObjectUpdateFlags

Mutable reference to this collision object’s update flags.

pub fn clear_update_flags(&mut self)

Clears the update flags of this collision object.

pub fn proxy_handle(&self) -> Option<BroadPhaseProxyHandle>

The collision object’s broad phase proxy unique identifier.

pub fn set_proxy_handle(&mut self, handle: Option<BroadPhaseProxyHandle>)

Set collision object’s broad phase proxy unique identifier.

pub fn position(&self) -> &Isometry<N>

The collision object position.

pub fn predicted_position(&self) -> Option<&Isometry<N>>

The predicted collision object position.

pub fn set_position(&mut self, pos: Isometry<N>)

Sets the position of the collision object and resets the predicted position to None.

pub fn set_position_with_prediction( &mut self, pos: Isometry<N>, prediction: Isometry<N>, )

Sets the position of the collision object and resets the predicted position.

pub fn set_predicted_position(&mut self, pos: Option<Isometry<N>>)

Sets the predicted position of the collision object.

pub fn set_deformations(&mut self, coords: &[N])

Deforms the underlying shape if possible.

Panics if the shape is not deformable.

pub fn shape(&self) -> &ShapeHandle<N>

The collision object shape.

pub fn set_shape(&mut self, shape: ShapeHandle<N>)

Set the collision object shape.

pub fn collision_groups(&self) -> &CollisionGroups

The collision groups of the collision object.

pub fn set_collision_groups(&mut self, groups: CollisionGroups)

Sets the collision groups of this collision object.

pub fn query_type(&self) -> GeometricQueryType<N>

The kind of queries this collision object is expected to .

pub fn set_query_type(&mut self, query_type: GeometricQueryType<N>)

Sets the GeometricQueryType of the collision object. Use CollisionWorld::set_query_type to use this method.

pub fn data(&self) -> &T

Reference to the user-defined data associated to this object.

pub fn data_mut(&mut self) -> &mut T

Mutable reference to the user-defined data associated to this object.