Trait ncollide3d::pipeline::object::CollisionObjectSet

pub trait CollisionObjectSet<N: RealField> {
    type CollisionObject: CollisionObjectRef<N>;
    type CollisionObjectHandle: CollisionObjectHandle;
    fn collision_object(
        &self, 
        handle: Self::CollisionObjectHandle
    ) -> Option<&Self::CollisionObject>;

    fn foreach(
        &self, 
        f: impl FnMut(Self::CollisionObjectHandle, &Self::CollisionObject)
    );
}

Trait implemented by sets of collision objects.

A set of collision object map a handle of type Self::CollisionObjectHandle with a collision object of type Self::CollisionObject.

Associated Types

type CollisionObject: CollisionObjectRef<N>

Type of the collision object stored into this set.

type CollisionObjectHandle: CollisionObjectHandle

Type of the handles identifying collision objects.

Loading content...
Required methods

fn collision_object(
&self,
handle: Self::CollisionObjectHandle
) -> Option<&Self::CollisionObject>

Gets the collision object identified by the given handle.

fn foreach(
&self,
f: impl FnMut(Self::CollisionObjectHandle, &Self::CollisionObject)
)

Applies a closure to every collision object (and their handle) stored into this set.

Loading content...