ncollide3d::procedural::path

Trait PolylineCompatibleCap

pub trait PolylineCompatibleCap<N: RealField + Copy> {
    // Required methods
    fn gen_start_cap(
        &self,
        attach_id: u32,
        pattern: &[Point3<N>],
        pt: &Point3<N>,
        dir: &Vector3<N>,
        closed: bool,
        coords: &mut Vec<Point3<N>>,
        indices: &mut Vec<Point3<u32>>,
    );
    fn gen_end_cap(
        &self,
        attach_id: u32,
        pattern: &[Point3<N>],
        pt: &Point3<N>,
        dir: &Vector3<N>,
        closed: bool,
        coords: &mut Vec<Point3<N>>,
        indices: &mut Vec<Point3<u32>>,
    );
}

Trait to be implemented by caps compatible with a PolylinePattern.

Required Methods

fn gen_start_cap( &self, attach_id: u32, pattern: &[Point3<N>], pt: &Point3<N>, dir: &Vector3<N>, closed: bool, coords: &mut Vec<Point3<N>>, indices: &mut Vec<Point3<u32>>, )

Generates the mesh for the cap at the beginning of a path.

fn gen_end_cap( &self, attach_id: u32, pattern: &[Point3<N>], pt: &Point3<N>, dir: &Vector3<N>, closed: bool, coords: &mut Vec<Point3<N>>, indices: &mut Vec<Point3<u32>>, )

Generates the mesh for the cap at the end of a path.