Module ncollide3d::query

Non-persistant pairwise geometric queries.

Modules

algorithms
Algorithms needed for distance and penetration depth computation.

closest_points_internal
Implementation details of the closest_points function.

contacts_internal
Implementation details of the contact and contacts functions.

distance_internal
Implementation details of the distance function.

point_internal
Point inclusion and projection.

proximity_internal
Implementation details of the proximity function.

ray_internal
Ray-casting related definitions and implementations.

time_of_impact_internal
Implementation details of the time_of_impact function.

Structs

Contact
Geometric description of a contact.

ContactKinematic
Local contact kinematic of a pair of solids around two given points.

ContactManifold
A contact manifold.

ContactPrediction
The prediction parameters for contact determination.n

PointInterferencesCollector
Bounding Volume Tree visitor collecting nodes that may contain a given point.

PointProjection
Description of the projection of a point on a shape.

Ray
A Ray.

RayInterferencesCollector
Bounding Volume Tree visitor collecting interferences with a given ray.

RayIntersection
Structure containing the result of a successful ray cast.

RayIntersectionCostFn
A search thet selects the objects that has the smallest time of impact with a given ray.

TrackedContact
A contact combined with contact kinematic information as well as a persistant identifier.

Enums

ClosestPoints
Closest points information.

Proximity
Proximity information.

Traits

PointQuery
Trait of objects that can be tested for point inclusion and projection.

PointQueryWithLocation
Returns shape-specific info in addition to generic projection information

RayCast
Traits of objects which can be transformed and tested for intersection with a ray.

Functions

closest_points
Computes the pair of closest points between two shapes.

contact
Computes one contact point between two shapes.

distance
Computes the minimum distance separating two shapes.

proximity
Tests whether two shapes are in intersecting or separated by a distance smaller than margin.

time_of_impact
Computes the smallest time of impact of two shapes under translational movement.