Struct ncollide3d::query::algorithms::epa3::EPA

pub struct EPA<N: RealField + Copy> { /* private fields */ }

The Expanding Polytope Algorithm in 3D.

Implementations

impl<N: RealField + Copy> EPA<N>

pub fn new() -> Self

Creates a new instance of the 3D Expanding Polytope Algorithm.

pub fn project_origin<G>(
&mut self,
m: &Isometry<N>,
g: &G,
simplex: &VoronoiSimplex<N>
) -> Option<Point<N>>where
G: SupportMap<N> + ?Sized,

Projects the origin on boundary of the given shape.

The origin is assumed to be inside of the shape. If it is outside use the GJK algorithm instead. Return None if the origin is not inside of the shape or if the EPA algorithm failed to compute the projection.

pub fn closest_points<G1, G2>(
&mut self,
m1: &Isometry<N>,
g1: &G1,
m2: &Isometry<N>,
g2: &G2,
simplex: &VoronoiSimplex<N>
) -> Option<(Point<N>, Point<N>, Unit<Vector<N>>)>where
G1: SupportMap<N> + ?Sized,
G2: SupportMap<N> + ?Sized,

Projects the origin on a shape unsing the EPA algorithm.

The origin is assumed to be located inside of the shape. Returns None if the EPA fails to converge or if g1 and g2 are not penetrating.