Module ncollide3d::shape

Collision shapes supported by ncollide.

Structs

Ball
A Ball shape.

Capsule
SupportMap description of a capsule shape with its principal axis aligned with the y axis.

ClippingCache
A cache used for polygonal clipping.

Compound
A compound shape with an aabb bounding volume.

Cone
SupportMap description of a cylinder shape with its principal axis aligned with the y axis.

ConvexHull
A convex polyhedron without degenerate faces.

ConvexPolygonalFeature
Represents a convex polygonal approximation of a face of a solid.

Cuboid
Shape of a box.

Cylinder
SupportMap description of a cylinder shape with its principal axis aligned with the y axis.

HeightField
An heightfield implicitly discretized with triangles.

HeightFieldCellStatus
The status of the cell of an heightfield.

Plane
SupportMap description of a plane.

Polyline
A polygonal line.

Segment
A segment shape.

ShapeHandle
A shared handle to an abstract shape.

Tetrahedron
A tetrahedron with 4 vertices.

TriMesh
A 3d triangle mesh.

TriMeshFace
A face of a triangle mesh.

Triangle
A triangle shape.

Enums

DeformationsType
The type of elements used to describe a deformation on a collision object.

FeatureId
An identifier of a feature of a convex polyhedron.

SegmentPointLocation
Logical description of the location of a point on a triangle.

TetrahedronPointLocation
Logical description of the location of a point on a triangle.

TrianglePointLocation
Description of the location of a point on a triangle.

Traits

CompositeShape
Trait implemented by shapes composed of multiple simpler shapes.

ConvexPolyhedron
Trait implemented by all convex polyhedron.

DeformableShape
Trait implemented by deformable shapes.

Shape
Trait implemented by all shapes supported by ncollide.

SupportMap
Traits of convex shapes representable by a support mapping function.