Struct ncollide3d::shape::ConvexPolygonalFeature

pub struct ConvexPolygonalFeature<N: RealField> {
    pub vertices: Vec<Point<N>>,
    pub edge_normals: Vec<Vector<N>>,
    pub normal: Option<Unit<Vector<N>>>,
    pub feature_id: FeatureId,
    pub vertices_id: Vec<FeatureId>,
    pub edges_id: Vec<FeatureId>,
}

Represents a convex polygonal approximation of a face of a solid.

It is never checked if the vertices actually form a convex polygon. If they do not, results of any geometric query may end up being invalid.

Fields
vertices: Vec<Point<N>>

The vertices of this face.

edge_normals: Vec<Vector<N>>

The outward normal of the edges if it is a face.

normal: Option<Unit<Vector<N>>>

The normal of this feature if it is a face.

feature_id: FeatureId

The shape-dependent identifier of this feature.

vertices_id: Vec<FeatureId>

The shape-dependent indentifier of each vertex of this feature.

edges_id: Vec<FeatureId>

The shape-dependent indentifier of each edge of this feature.

Implementations

impl<N: RealField> ConvexPolygonalFeature<N>

pub fn new() -> Self

Creates a new empty convex polygonal faces.

pub fn with_size(n: usize) -> Self

Creates a new convex polygonal feature with all field initialized with n zero elements.

pub fn clear(&mut self)

Removes all the vertices, normals, and feature IDs of this feature.

pub fn transform_by(&mut self, m: &Isometry<N>)

Transforms all the vertices and normals of this feature by the given isometry.

pub fn push(&mut self, pt: Point<N>, id: FeatureId)

Adds a vertex to this face.

It is not checked whether pt breaks the convexity of the polyhedral face.

pub fn nvertices(&self) -> usize

The number of vertices of this face.

pub fn vertices(&self) -> &[Point<N>]

The vertices of this convex polygonal face.

pub fn nedges(&self) -> usize

The number of edges of this convex polygonal face.

pub fn edge(&self, edge_id: FeatureId) -> Option<Segment<N>>

Retrieves the edge with the given feature id.

pub fn push_scaled_edge_normal(&mut self, normal: Vector<N>)

Adds a scaled edge normal to this face.

pub fn push_edge_normal(&mut self, normal: Unit<Vector<N>>)

Adds an edge normal to this face.

pub fn recompute_edge_normals(&mut self)

Automatically recomputes the scaled edge normals (3D only).

Panics if the ambient space is not 3D.

pub fn project_point(&self, pt: &Point<N>) -> Option<Contact<N>>

Transforms all the vertices of this feature by the given isometry.

pub fn set_normal(&mut self, normal: Unit<Vector<N>>)

Sets the outward normal of this convex polygonal face.

pub fn push_edge_feature_id(&mut self, id: FeatureId)

Add the shape-dependent identifier of a edge of this feature (if it is a face).

pub fn set_feature_id(&mut self, id: FeatureId)

Add the shape-dependent identifier of this feature.

pub fn clip(
&self,
other: &Self,
normal: &Unit<Vector<N>>,
prediction: &ContactPrediction<N>,
cache: &mut ClippingCache<N>,
out: &mut Vec<(Contact<N>, FeatureId, FeatureId)>
)

Generate contacts between self and other using polygonal clipping, iif. they both have at least three vertices.

If either self or other has less than three vertices, this does nothing.

pub fn add_contact_to_manifold(
&self,
other: &Self,
c: Contact<N>,
m1: &Isometry<N>,
f1: FeatureId,
proc1: Option<&dyn ContactPreprocessor<N>>,
m2: &Isometry<N>,
f2: FeatureId,
proc2: Option<&dyn ContactPreprocessor<N>>,
manifold: &mut ContactManifold<N>
)

Given a contact between two polygonal features, adds it to a contact manifold.