ncollide3d::shape

Trait DeformableShape

pub trait DeformableShape<N: RealField + Copy> {
    // Required methods
    fn deformations_type(&self) -> DeformationsType;
    fn set_deformations(&mut self, coords: &[N]);
    fn update_local_approximation(
        &self,
        coords: &[N],
        approx: &mut LocalShapeApproximation<N>,
    );
}

Trait implemented by deformable shapes.

Required Methods

fn deformations_type(&self) -> DeformationsType

The type of degrees of freedom this shape has.

fn set_deformations(&mut self, coords: &[N])

Updates some the degrees of freedom of this shape. Updates all the degrees of freedom of this shape.

fn update_local_approximation( &self, coords: &[N], approx: &mut LocalShapeApproximation<N>, )

Updates the given local approximation of this shape.