Trait ncollide3d::shape::Shape

pub trait Shape<N: Real>: Send + Sync + Any + GetTypeId {
    fn aabb(&self, m: &Isometry<N>) -> AABB<N>;

    fn bounding_sphere(&self, m: &Isometry<N>) -> BoundingSphere<N> { ... }

    fn subshape_transform(&self, _: usize) -> Option<Isometry<N>> { ... }

    fn as_ray_cast(&self) -> Option<&RayCast<N>> { ... }

    fn as_point_query(&self) -> Option<&PointQuery<N>> { ... }

    fn as_convex_polyhedron(&self) -> Option<&ConvexPolyhedron<N>> { ... }

    fn as_support_map(&self) -> Option<&SupportMap<N>> { ... }

    fn as_composite_shape(&self) -> Option<&CompositeShape<N>> { ... }

    fn is_convex_polyhedron(&self) -> bool { ... }

    fn is_support_map(&self) -> bool { ... }

    fn is_composite_shape(&self) -> bool { ... }
}

Trait implemented by all shapes supported by ncollide.

This allows dynamic inspection of the shape capabilities.

Required Methods

fn aabb(&self, m: &Isometry<N>) -> AABB<N>

The AABB of self.

Provided Methods

fn bounding_sphere(&self, m: &Isometry<N>) -> BoundingSphere<N>

The bounding sphere of self.

fn subshape_transform(&self, _: usize) -> Option<Isometry<N>>

The transform of a specific subshape.

Return None if the transform is known to be the identity.

fn as_ray_cast(&self) -> Option<&RayCast<N>>

The RayCast implementation of self.

fn as_point_query(&self) -> Option<&PointQuery<N>>

The PointQuery implementation of self.

fn as_convex_polyhedron(&self) -> Option<&ConvexPolyhedron<N>>

The convex polyhedron representation of self if applicable.

fn as_support_map(&self) -> Option<&SupportMap<N>>

The support mapping of self if applicable.

fn as_composite_shape(&self) -> Option<&CompositeShape<N>>

The composite shape representation of self if applicable.

fn is_convex_polyhedron(&self) -> bool

Whether self uses a conve polyhedron representation.

fn is_support_map(&self) -> bool

Whether self uses a supportmapping-based representation.

fn is_composite_shape(&self) -> bool

Whether self uses a composite shape-based representation.

Methods

impl<N: Real> Shape<N>

Trait for casting shapes to its exact represetation.

pub fn is_shape<T: Shape<N>>(&self) -> bool

Tests if this shape has a specific type T.

pub fn as_shape<T: Shape<N>>(&self) -> Option<&T>

Performs the cast.