Trait ncollide3d::shape::SupportMap

pub trait SupportMap<N: Real> {
    fn support_point(
        &self, 
        transform: &Isometry<N>, 
        dir: &Vector<N>
    ) -> Point<N>;

    fn support_point_toward(
        &self, 
        transform: &Isometry<N>, 
        dir: &Unit<Vector<N>>
    ) -> Point<N> { ... }
}

Traits of convex shapes representable by a support mapping function.

Parameters:

V - type of the support mapping direction argument and of the returned point.

Required Methods

fn support_point(&self, transform: &Isometry<N>, dir: &Vector<N>) -> Point<N>

Evaluates the support function of the object.

A support function is a function associating a vector to the shape point which maximizes their dot product.

Provided Methods

fn support_point_toward(
&self,
transform: &Isometry<N>,
dir: &Unit<Vector<N>>
) -> Point<N>

Same as self.support_point except that dir is normalized.