Trait ncollide3d::shape::SupportMap

pub trait SupportMap<N: RealField> {
    fn local_support_point(&self, dir: &Vector<N>) -> Point<N>;

    fn local_support_point_toward(&self, dir: &Unit<Vector<N>>) -> Point<N> { ... }

    fn support_point(
        &self, 
        transform: &Isometry<N>, 
        dir: &Vector<N>
    ) -> Point<N> { ... }

    fn support_point_toward(
        &self, 
        transform: &Isometry<N>, 
        dir: &Unit<Vector<N>>
    ) -> Point<N> { ... }
}

Traits of convex shapes representable by a support mapping function.

Parameters:

V - type of the support mapping direction argument and of the returned point.

Required methods

fn local_support_point(&self, dir: &Vector<N>) -> Point<N>

Loading content...
Provided methods

fn local_support_point_toward(&self, dir: &Unit<Vector<N>>) -> Point<N>

Same as self.local_support_point except that dir is normalized.

fn support_point(&self, transform: &Isometry<N>, dir: &Vector<N>) -> Point<N>

fn support_point_toward(
&self,
transform: &Isometry<N>,
dir: &Unit<Vector<N>>
) -> Point<N>

Same as self.support_point except that dir is normalized.

Loading content...