Trait ncollide3d::utils::IsometryOps

pub trait IsometryOps<N: Real> {
    fn absolute_transform_vector(&self, v: &Vector<N>) -> Vector<N>;

    fn inverse_transform_vector(&self, v: &Vector<N>) -> Vector<N>;

    fn inverse_transform_point(&self, p: &Point<N>) -> Point<N>;

    fn inverse_transform_unit_vector(
        &self, 
        v: &Unit<Vector<N>>
    ) -> Unit<Vector<N>> { ... }
}

Extra operations with isometries.

Required Methods

fn absolute_transform_vector(&self, v: &Vector<N>) -> Vector<N>

Transform a vector by the absolute value of the homogeneous matrix equivalent to self.

fn inverse_transform_vector(&self, v: &Vector<N>) -> Vector<N>

Transform a vector by the inverse of self.

fn inverse_transform_point(&self, p: &Point<N>) -> Point<N>

Transform a point by the inverse of self.

Provided Methods

fn inverse_transform_unit_vector(&self, v: &Unit<Vector<N>>) -> Unit<Vector<N>>

Transform a unit vector by the inverse of self.